refactor menu serviceto get meal based on the date
