+++ back/controller/MenuController.js
@@ -2,7 +2,8 @@ import MenusServices from '../services/MenusServices';
 
 export default class MenuController {
   static getMenu(req, res) {
-    res.status(200).send(MenusServices.getMenu());
+    const { date } = req.query;
+    res.status(200).send(MenusServices.getMenu(date));
   }
 
   static setupMenu(req, res) {
