--- back/services/TestServices.js
@@ -1,9 +1,0 @@
-export default class TestServices {
-  constructor() {
-    this.test = 'test data';
-  }
-
-  getTest() {
-    return this.test;
-  }
-}
